move GeneralizedCoastlineCheck from linear.edges to linear.lines
