+++ src/main/java/org/openstreetmap/atlas/checks/validation/linear/lines/GeneralizedCoastlineCheck.java
@@ -1,4 +1,4 @@
-package org.openstreetmap.atlas.checks.validation.linear.edges;
+package org.openstreetmap.atlas.checks.validation.linear.lines;
 
 import java.util.Arrays;
 import java.util.List;
